+++ package.json
@@ -11,7 +11,7 @@
     "private::build:js-dev": "run-s \"private::build -- --mode development\"",
     "private::build:js-test": "run-s \"private::build -- --config webpack.test.config.js\"",
     "private::build:js-test-watch": "run-s \"private::build -- --config webpack.test.config.js --watch\"",
-    "private::build:extract-meta": "node ./extract-meta src/dash-table/DataTable.js > dash_table/metadata.json",
+    "private::build:extract-meta": "node ./extract-meta src/dash-table/dash/DataTable.js > dash_table/metadata.json",
     "private::build:copy-package-info": "cp package.json dash_table/package-info.json",
     "private::build:generate-classes": "python -c \"import dash; dash.development.component_loader.generate_classes('dash_table', 'dash_table/metadata.json')\"",
     "private::build:py": "run-s private::build:copy-package-info private::build:extract-meta private::build:generate-classes",
